Add export test cases
